% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_cells.R
\name{annotate_cells}
\alias{annotate_cells}
\title{annotate_cells.R}
\usage{
annotate_cells(
  seurat_obj = NULL,
  markers = "CellMarker2.0",
  plot_directory = NULL,
  CellMarker_path = NULL,
  celltypes = NULL,
  tissue = "normal peripheral blood",
  scina_probability_threshold = 0.5
)
}
\arguments{
\item{seurat_obj}{Input seurat object created using prepare_annotation_data()}

\item{markers}{Either "CellMarker2.0" (default) or named list of markers for cell types}

\item{plot_directory}{Output directory to save plots}

\item{CellMarker_path}{Path to CellMarker2.0}

\item{celltypes}{Can be used to limit the set of possible celltypes to be considered}

\item{tissue}{Only required if using CellMarker2.0}

\item{scina_probability_threshold}{Only keep celltype assignments with probability above this thresholds (Default 0.5)}
}
\value{
Seurat object
}
\description{
This function automatically annotates individual cells in a seurat object
}
