% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_clusters.R
\name{annotate_clusters}
\alias{annotate_clusters}
\title{annotate_clusters.R}
\usage{
annotate_clusters(
  seurat_obj = NULL,
  resolution = NULL,
  annotations = NULL,
  plot_directory = NULL
)
}
\arguments{
\item{seurat_obj}{Input seurat object}

\item{resolution}{Resolution to be used for clustering}

\item{annotations}{Named vector of cell types for each cluster}

\item{plot_directory}{Location to save final annotations UMAP}
}
\value{
Seurat object
}
\description{
This function adds manual celltype annotations to a seurat object
}
