% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_mutants.R
\name{annotate_mutants}
\alias{annotate_mutants}
\title{annotate_mutants.R}
\usage{
annotate_mutants(
  seurat_obj,
  variant,
  name = NULL,
  combine_strategy = "intersect",
  genotypes_file,
  reduction = "umap",
  plot_directory = NULL
)
}
\arguments{
\item{seurat_obj}{Input seurate object}

\item{variant}{Variant/s to be annotated}

\item{name}{Identifier given to annotated cells carrying specified variants}

\item{combine_strategy}{How cells are chosen when multiple variants are provided, either "intersect" (default) or "union"}

\item{genotypes_file}{Path to the *_genotypes.tsv file containing cell-specific variant info}

\item{reduction}{Which reduction to use when generating UMAP plots. Usually either
"umap" for ADT-base clusters (default) or "umap.variants" for variant-based clusters}

\item{plot_directory}{Location to save output plot}
}
\value{
Seurat object
}
\description{
This function adds manual variant annotations to a seurat object
}
