% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_variants.R
\name{cluster_variants}
\alias{cluster_variants}
\title{cluster_variants.R}
\usage{
cluster_variants(
  seurat_obj = NULL,
  genotypes_file,
  summary_file,
  plot_directory = NULL,
  n_features = 100,
  sample_ID = NULL,
  n_neighbors = 30,
  min_dist = 0.3,
  dist_metric = "euclidean",
  dbscan_eps = 0.8,
  min_data_portion = 0.8,
  rare_frequency_max = 0.02,
  min_overlap = 10L,
  min_alt_count = 5,
  prune_correlated_variants = FALSE,
  rare_var_weighting = 0.1,
  minPts = NULL,
  clustering_seed = NULL
)
}
\arguments{
\item{seurat_obj}{Optional input seurat object created using prepare_annotation_data()}

\item{genotypes_file}{Path to *genotypes.tsv file generated by find_somatic_variants()}

\item{summary_file}{Path to *summary.tsv file generated by find_somatic_variants()}

\item{plot_directory}{Optional path to a directory where a UMAP plot will be saved}

\item{n_features}{Target number of variants to use for clustering (Default = 100)}

\item{sample_ID}{Sample identifier to be added to seurat object}

\item{n_neighbors}{Passed to uwot::umap (Default = 30)}

\item{min_dist}{Passed to uwot::umap (Default = 0.3)}

\item{dist_metric}{Passed to uwot::umap (Default = "euclidean")}

\item{dbscan_eps}{Eps metric passed to dbscan::dbscan (Default = 0.8)}

\item{min_data_portion}{Minimum data availability for filtering variants (Default = 0.8)}

\item{rare_frequency_max}{Max frequency threshold for selecting rare variants (Default = 0.02)}

\item{min_overlap}{Minimum number of joint non-NA allele frequency values
shared between two cells for calulcating cosine distance (Default = 10)}

\item{min_alt_count}{Minimum number of cells carrying a mutation for it to be included for clustering (Default = 5)}

\item{prune_correlated_variants}{Filter out selected variants highly correlated with each other (Experimental) (Default = FALSE)}

\item{rare_var_weighting}{Proportion of clustering features that should be rare variants (Default = 0.1)}

\item{minPts}{minPts metric passed to dbscan::dbscan (Default = NULL)}

\item{clustering_seed}{Seed used for DBSCAN clustering (Default = NULL)}
}
\value{
Seurat object
}
\description{
This function performs cell clustering based on variant genotypes
}
