% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_QQ_plot.R
\name{generate_qq_plot}
\alias{generate_qq_plot}
\title{generate_QQ_plot.R}
\usage{
generate_qq_plot(
  results_dir = NULL,
  prefix = NULL,
  plot_dir,
  results_files = NULL
)
}
\arguments{
\item{results_dir}{The directory where cell type variant enrichment results
are stored, namely \code{{prefix}}_\code{{cell_type}}.tsv files}

\item{prefix}{Prefix used in \code{find_somatic_variants()} to generate results}

\item{plot_dir}{Directory to output QQ plot}

\item{results_files}{Optionally a vector of specific results files to use}
}
\value{
List of plots
}
\description{
Generates a QQ plot and histograms of p values from presumed null variants based
on VEP variant annotations.
}
