% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_annotation_data.R
\name{prepare_annotation_data}
\alias{prepare_annotation_data}
\title{prepare_annotation_data.R}
\usage{
prepare_annotation_data(
  input_file = NULL,
  sample_ID = NULL,
  cluster_resolutions = c(0.2, 0.35, 0.5, 0.8, 1, 1.2),
  save_directory = NULL,
  qc_plot_directory = NULL,
  variance_quantile = 0.25,
  neighbor_metric = "euclidean",
  k_param = 50
)
}
\arguments{
\item{input_file}{Path to input .dna+protein.h5 file}

\item{sample_ID}{Sample ID, or NULL to derive from .h5 filname (default)}

\item{cluster_resolutions}{Numeric vector of resolutions for Seurat::FindClusters (default: c(0.2, 0.35, 0.5, 0.8, 1, 1.2))}

\item{save_directory}{Output directory to save seurat object as .RDS, or NULL to prevent saving (default)}

\item{qc_plot_directory}{Directory to save QC plots to}

\item{variance_quantile}{Threshold for selecting the top quantile of ADT markers based on their variance}

\item{neighbor_metric}{Distance metric to be used for cell clustering by Seurat::RunUMAP and Seurat::FindNeighbors (eg. euclidean (default), cosine, correlation)}

\item{k_param}{k.param for Seurat::FindNeighbors, also used for n.neighbors in Seurat::RunUMAP}
}
\value{
Seurat object
}
\description{
This function loads annotation data from a tapestri .h5 file and generates a seurat object ready for annotation
}
