% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vep_functions.R
\name{run_vep}
\alias{run_vep}
\title{Run VEP and return either the output path, a data.frame, or a VCF object}
\usage{
run_vep(
  input,
  output = tempfile(fileext = ".vep.vcf"),
  return = c("path", "data.frame", "vcf"),
  args = character()
)
}
\arguments{
\item{input}{Path to VEP input file}

\item{output}{desired output path (.vcf recommended)}

\item{return}{Type of object that the function should return, one of "path", "data.frame", or "vcf"}

\item{args}{extra VEP CLI args (character vector)}
}
\description{
Run VEP and return either the output path, a data.frame, or a VCF object
}
