% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vep_install_cache.R
\name{vep_install_cache}
\alias{vep_install_cache}
\title{Install/populate a VEP cache using your chosen backend
Works with singularity or system/conda. Run on a node with internet.}
\usage{
vep_install_cache(
  species = .get_cfg()$species,
  assembly = .get_cfg()$assembly,
  cache_version = NULL,
  plugins = TRUE,
  fasta = TRUE
)
}
\arguments{
\item{species}{e.g. "homo_sapiens"}

\item{assembly}{e.g. "GRCh38" or "GRCh37"}

\item{cache_version}{optional Ensembl cache version (e.g. 114)}

\item{plugins}{also fetch Plugins/}

\item{fasta}{also fetch reference FASTA}
}
\description{
Install/populate a VEP cache using your chosen backend
Works with singularity or system/conda. Run on a node with internet.
}
